+++ dit/profiles/__init__.py
@@ -1,0 +1,10 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+This module contains various information ``profiles'', decomposing the total information into the information contained at different scales.
+"""
+
+from complexity_profile import ComplexityProfile
+from marginal_utility_of_information import MUIProfile
+from schneidman import SchneidmanProfile